refactor(dispatch): split the fire-path core into focused modules (Phase B)
Finishes the dispatch.ts split. The remaining fire-path core moves out of dispatch.ts into
sibling modules; dispatch.ts becomes a 29-line public barrel that re-exports the surface so
every @/core/dispatch consumer (server.ts namespace import) + the 3 mock-the-path test files
are unchanged.

Modules: dispatch-state.ts (the .dispatch-state.json layer — SOLE owner of the stateQueue
mutex), dispatch-turns.ts (fireDispatchTurn + the inFlightTurns/pendingLadderRedispatches
singletons + concurrencyGate + budget + scheduleLadderRedispatch helper), dispatch-session-
death.ts (the recovery ladder + reconciliation), dispatch-cycle.ts (the dispatching/timer
singletons + dispatchTasks loop + start/stop/getDispatchInfo), dispatch-single.ts
(dispatchSingleTask), dispatch-workflow.ts (dispatchWorkflowTask + buildWorkflowDispatchMessage).

Hazards handled:
- ONE mutex: stateQueue lives only in dispatch-state.ts; every caller imports withStateLock
  from there (verified by grep). A split mutex would silently double-write the state file.
- The single→turns→session-death→single cycle is broken with a LAZY `import('./dispatch-single')`
  inside handleSessionDeath's delayed re-dispatch — no static cycle and no dependency on
  barrel-load wiring (replaces an earlier setRedispatch DI that would have thrown "not wired"
  for any non-barrel submodule importer).
- Each remaining singleton lives in exactly one module; turns↔session-death is a runtime-only
  cycle (accesses inside function bodies). The .dispatch-state.json path/shape (a cross-package
  ledger contract) is unchanged. Logger pretty-suppress allowlist gains the new module names.

Pure relocation, no behavior change, NO dedup (prepareAndFireRegularDispatch is a separate
follow-up). Fixed a brittle source-scanning test that read dispatch.ts for buildWorkflowDispatchMessage
(now in dispatch-workflow.ts) + the no-hardcoded-roster scan (now over the prompt modules).

Verified HERE: typecheck + lint clean; full suite 5106/0; full 3-target binary build; docker
boot + create-task smoke (the split fire-core runs, no module-wiring crash). The authoritative
exactly-once gate is the bare-metal live-OpenClaw stress plan in
tasks/dispatch-phase-b-handoff.md (burst concurrency / session-death ladder / restart recovery /
soak — exactly-once checked via ledger + audit), to be run before merge.
