+++ packages/core/src/logger.ts
@@ -188,6 +188,12 @@ function suppressPrettyInfo(entry: LogEntry): boolean {
     'hot-reload-coordinator',
     'mcporter',
     'dispatch',
+    'dispatch-state',
+    'dispatch-turns',
+    'dispatch-session-death',
+    'dispatch-workflow',
+    'dispatch-cycle',
+    'dispatch-single',
     'watchdog',
     'doctor',
     'lifecycle',
