+++ src/core/dispatch.ts
@@ -1,1588 +1,29 @@
 /**
- * Task dispatch system for Bakin.
- * Periodically checks for TODO tasks and dispatches them to agents via the runtime adapter.
+ * Task dispatch system for Bakin — public barrel.
+ *
+ * The implementation is split across sibling modules (dispatch-types,
+ * dispatch-failures, dispatch-state, dispatch-board, dispatch-context-blocks,
+ * dispatch-prompts, dispatch-turns, dispatch-session-death, dispatch-workflow,
+ * dispatch-cycle, dispatch-single). This file re-exports the public surface so
+ * `@/core/dispatch` consumers (server.ts namespace import; the test files that
+ * mock.module the path) keep their import path unchanged. Internal modules
+ * import each other directly — never through this barrel — to avoid self-cycles.
  */
-import { writeFileSync } from 'fs'
-import { join } from 'path'
-import { createLogger } from './logger'
-import { getSettings } from './settings'
-import { appendAudit } from './audit'
-import { recordUsage } from './usage'
-import { getAppServices } from './app-services'
-import { getRuntimeMainAgentId, RuntimeTurnError, type MessageResult } from '@bakin/core/adapters/runtime'
-import { claimNextRun, currentSeq, loseRun, settleRun, spendTotal, type ClaimNextRunResult } from './execution-ledger'
-import { meterAgentTurn } from './agent-cost'
-import { resolveTurnModel, type ResolvedTurn, type RoutingConfig } from './model-routing'
-import { evaluateBudget, dayStartMs, monthStartMs, type BudgetPolicy, type BudgetSpend, type BudgetDecision } from './budget'
-import { getBootId } from './boot-id'
-import { getHookRegistry } from '@bakin/core/hooks/hook-registry-singleton'
-import {
-  buildTaskLessonQuery,
-} from './agent-packages/lesson-retrieval'
-import {
-  blockTask as blockStoredTask,
-  moveTask as moveStoredTask,
-  readTaskboard,
-} from './task-store'
-import {
-  formatDispatchError,
-  classifyDispatchError,
-  classifyDispatchFailureDetail,
-  formatSanitizedRuntimeFailure,
-} from './dispatch-failures'
+import { dispatchSingleTask } from './dispatch-single'
 
-// Re-export the pure failure-classification surface so `@/core/dispatch`
-// consumers (and the dispatch-error-classification test) keep their import path.
-export { classifyDispatchError, classifyDispatchFailureDetail, formatSanitizedRuntimeFailure }
+// ── Public surface (unchanged for every `@/core/dispatch` consumer) ──
+export { classifyDispatchError, classifyDispatchFailureDetail, formatSanitizedRuntimeFailure } from './dispatch-failures'
+export { loadDispatchState, clearDispatchMarker } from './dispatch-state'
+export { isTaskDispatchEligible } from './dispatch-board'
+export { __resetLessonBlockCache, buildDispatchLessonBlock, buildDispatchAssetBlock } from './dispatch-context-blocks'
+export { buildDispatchMessage } from './dispatch-prompts'
+export { budgetGate, getInFlightTurnCount, awaitDispatchIdle } from './dispatch-turns'
+export { dispatchTasks, start, stop, getDispatchInfo } from './dispatch-cycle'
+export { dispatchSingleTask }
 
-import type {
-  DispatchTask,
-  DispatchColumns,
-  DispatchEligibility,
-  DispatchIneligibleReason,
-  SessionDeathDiagnosisLite,
-  SessionDeathState,
-  DispatchState,
-  DispatchContinuationContext,
-  DispatchRosterAgent,
-  InFlightTurn,
-  ConcurrencyGate,
-} from './dispatch-types'
-import {
-  readDispatchColumns,
-  isTaskDispatchEligible,
-  findDispatchTaskSnapshot,
-  taskAlreadyLeftActiveWork,
-  shouldBlockAfterDispatchFailure,
-  addTaskLog,
-  moveTaskToInProgress,
-  tryAddTaskLog,
-} from './dispatch-board'
-import {
-  __resetLessonBlockCache,
-  buildDispatchLessonBlock,
-  buildDispatchAssetBlock,
-} from './dispatch-context-blocks'
-import {
-  mcporterHelpers,
-  sharedExecutionToolDocs,
-  outputDisciplineSection,
-  buildCorrectiveSection,
-  buildDecompositionMessage,
-  buildDispatchMessage,
-} from './dispatch-prompts'
-import {
-  withStateLock,
-  getFailureRecord,
-  cooldownForFailure,
-  getDispatchMarkerTaskId,
-  removeDispatchMarkersForTask,
-  trimDispatched,
-  loadDispatchState,
-  saveDispatchState,
-  clearDispatchMarker,
-} from './dispatch-state'
-
-// Re-export the moved public surface so `@/core/dispatch` consumers + tests
-// keep their import paths.
 export type {
   DispatchEligibility,
   DispatchIneligibleReason,
   DispatchContinuationContext,
   DispatchRosterAgent,
-}
-export {
-  isTaskDispatchEligible,
-  buildDispatchAssetBlock,
-  buildDispatchLessonBlock,
-  __resetLessonBlockCache,
-  buildDispatchMessage,
-  loadDispatchState,
-  clearDispatchMarker,
-}
-
-const log = createLogger('dispatch')
-const hooks = () => getHookRegistry()
-
-/**
- * Task-work sends get a per-dispatch session: a fresh, deterministic
- * threadId per attempt so context can't accumulate across tasks, forensics
- * knows exactly which provider session to inspect, and a corrective
- * re-dispatch never replays a dead session's bloated context. Notification
- * sends (orchestrator/watchdog/doctor) deliberately stay in the agent's
- * default session and don't go through here.
- */
-async function sendDispatchMessage(agentId: string, content: string, threadId: string, routing?: ResolvedTurn): Promise<MessageResult> {
-  return getAppServices().runtime.messaging.send({
-    agentId,
-    content,
-    threadId,
-    ...(routing?.model ? { model: routing.model } : {}),
-    ...(routing?.thinking ? { thinking: routing.thinking } : {}),
-    metadata: { oversizedOutputBytes: getSettings().dispatch.oversizedOutputBytes },
-  })
-}
-
-// Budget warn/defer audits fire once per (event, scope, window, window-start)
-// so a cap sitting at 85% — or a deferred task re-evaluated every cycle —
-// doesn't spam the audit log. Keyed by window start, so it auto-rolls when
-// the day/month turns over. In-memory; resets on restart (acceptable).
-const budgetAuditedWindows = new Set<string>()
-
-function auditBudgetOnce(contentDir: string, event: 'budget.warn' | 'budget.deferred', agent: string, detail: Record<string, unknown>, windowStartMs: number): void {
-  const key = `${event}:${detail.scope}:${detail.window}:${windowStartMs}`
-  if (budgetAuditedWindows.has(key)) return
-  budgetAuditedWindows.add(key)
-  appendAudit(contentDir, event, agent, detail)
-}
-
-/**
- * Consult the budget policy before a dispatch claims a run. allow / warn /
- * defer; warn and defer are audited (debounced per window). FAIL-CLOSED: if
- * the spend ledger can't be read, defer — consistent with the ledger's
- * existing fail-closed posture. No policy (plugin absent / empty) → allow.
- */
-export async function budgetGate(agentId: string, contentDir: string, spendCache?: Map<string, number>): Promise<BudgetDecision> {
-  let policy: BudgetPolicy | undefined
-  try {
-    policy = (await hooks().invoke<BudgetPolicy>('models.getBudgetPolicy', {})) ?? undefined
-  } catch (err) {
-    log.error('Budget policy read failed; allowing (no policy)', err, { agentId })
-    return { action: 'allow' }
-  }
-  if (!policy || (!policy.global && !policy.perAgent)) return { action: 'allow' }
-
-  const now = Date.now()
-  const dayStart = dayStartMs(now)
-  const monthStart = monthStartMs(now)
-  // Memoize spend reads within a dispatch cycle: the two GLOBAL totals are
-  // identical for every task in the cycle, and costs are only recorded on
-  // settle (after the loop), so a per-cycle cache is safe and saves ~2 SQL
-  // aggregates per task. Cache key includes the window start so it can't
-  // bleed across a day/month boundary.
-  const spendAt = (agent: string | undefined, sinceMs: number): number => {
-    const key = `${agent ?? '*'}:${sinceMs}`
-    const hit = spendCache?.get(key)
-    if (hit !== undefined) return hit
-    const v = spendTotal(agent ? { agent, sinceMs } : { sinceMs })
-    spendCache?.set(key, v)
-    return v
-  }
-  let spend: BudgetSpend
-  try {
-    spend = {
-      globalDayMicros: spendAt(undefined, dayStart),
-      globalMonthMicros: spendAt(undefined, monthStart),
-      agentDayMicros: spendAt(agentId, dayStart),
-      agentMonthMicros: spendAt(agentId, monthStart),
-    }
-  } catch (err) {
-    log.error('Budget spend read failed; deferring (fail-closed)', err, { agentId })
-    auditBudgetOnce(contentDir, 'budget.deferred', agentId, { scope: 'global', window: 'daily', reason: 'ledger-unavailable' }, dayStart)
-    return { action: 'defer', scope: 'global', window: 'daily', spentUsdMicros: 0, capUsdMicros: 0 }
-  }
-
-  const decision = evaluateBudget({ policy, agent: agentId, spend })
-  if (decision.action !== 'allow') {
-    const windowStart = decision.window === 'monthly' ? monthStart : dayStart
-    const event = decision.action === 'defer' ? 'budget.deferred' : 'budget.warn'
-    auditBudgetOnce(contentDir, event, agentId, {
-      scope: decision.scope, window: decision.window,
-      spentUsdMicros: decision.spentUsdMicros, capUsdMicros: decision.capUsdMicros,
-    }, windowStart)
-  }
-  return decision
-}
-
-/** True when budget says defer — the shared shape all three dispatch paths use. */
-async function deferForBudget(agentId: string, contentDir: string, spendCache?: Map<string, number>): Promise<boolean> {
-  return (await budgetGate(agentId, contentDir, spendCache)).action === 'defer'
-}
-
-/**
- * Resolve the per-turn model/thinking for a dispatch from the Bakin-owned
- * routing policy (stored in the models plugin; read via hook). Returns {} —
- * inherit, unchanged dispatch — when no plugin/config/match applies or the
- * read fails. Never throws into the dispatch path.
- */
-async function resolveDispatchRouting(task: DispatchTask, isRecovery: boolean): Promise<ResolvedTurn> {
-  try {
-    const config = await hooks().invoke<RoutingConfig>('models.getRoutingConfig', {})
-    if (!config) return {}
-    return resolveTurnModel({
-      task: { tags: task.tags, scheduleJobId: task.scheduleJobId, workflowId: task.workflowId, parentId: task.parentId },
-      isRecovery,
-      config,
-    })
-  } catch (err) {
-    log.error('Routing resolve failed; using agent default', err, { id: task.id })
-    return {}
-  }
-}
-
-/**
- * Record the cost of a settled turn. The threadId IS the ledger run id, so
- * the row is first-write-wins idempotent. Pricing is delegated to the models
- * plugin via the `models.priceTurn` hook (core stays pricing-agnostic);
- * absent plugin → null model/cost, tokens still recorded ("unmetered"). The
- * same data also feeds the live usage recorder. Never throws into the settle
- * path — a metering failure must not fail a successful turn.
- */
-function recordTurnCost(runId: string, taskId: string, agent: string, result: MessageResult, resolvedModel?: string): Promise<void> {
-  return meterAgentTurn({ runId, taskId, agent, result, resolvedModel })
-}
-
-/**
- * Atomically mint the next per-attempt session key AND claim the live-run
- * slot in the execution ledger. The runs table's partial unique index (one
- * running row per exec key) is the correctness mechanism: a duplicate
- * dispatch (overlapping cycle, manual kick racing the cycle, second server
- * process) fails the INSERT instead of racing in-memory markers.
- *
- * `seq` stays a monotonic per-task counter (now ledger-owned, seeded from
- * the legacy .dispatch-state.json watermarks) — it must never be derived
- * from the failure count and reusing one could re-enter a live provider
- * session. Workflow steps carry the stepId so parallel step agents can't
- * collide: their exec key is `${taskId}:${stepId}`.
- *
- * Callers MUST settle the claim on every outcome (fireDispatchTurn's settle
- * handlers do) or release it via loseRun when the turn never fires.
- */
-function claimDispatchRun(taskId: string, targetAgent: string, stepId?: string): ClaimNextRunResult {
-  return claimNextRun({
-    taskId,
-    execKey: stepId ? `${taskId}:${stepId}` : taskId,
-    agent: targetAgent,
-    bootId: getBootId(),
-    runIdFor: (seq) => (stepId ? `task:${taskId}:step:${stepId}:d${seq}` : `task:${taskId}:d${seq}`),
-  })
-}
-
-/** Suppressed-duplicate bookkeeping shared by all three dispatch paths. */
-function auditDispatchSuppressed(contentDir: string, task: { id: string; title: string }, targetAgent: string, liveRunId: string | undefined, path: 'cycle' | 'single' | 'workflow'): void {
-  appendAudit(contentDir, 'task.dispatch_suppressed', targetAgent, {
-    id: task.id,
-    title: task.title,
-    liveRunId: liveRunId ?? null,
-    path,
-  })
-  log.warn('Dispatch suppressed — live run already exists', { id: task.id, liveRunId, path })
-}
-
-let dispatching = false
-let dispatchStartedAt = 0
-let dispatchTimer: NodeJS.Timeout | null = null
-const DISPATCH_TIMEOUT_MS = 3 * 60 * 1000 // 3 minutes max per dispatch cycle
-
-// Async mutex for .dispatch-state.json — serializes all reads/writes
-// How long after a session death before the automatic ladder re-dispatch
-// fires. Just enough to let the state lock release; deliberately NOT a
-// cooldown — retrying a deterministic failure later doesn't make it pass,
-// changing the approach (corrective prompt / decomposition) does.
-const SESSION_DEATH_REDISPATCH_DELAY_MS = 50
-
-// Ladder caps: regular tasks get corrective → decomposition → block;
-// workflow steps get corrective → block (step structure is owned by the
-// workflow engine — an agent must not decompose someone else's workflow).
-const MAX_SESSION_DEATHS_REGULAR = 3
-const MAX_SESSION_DEATHS_WORKFLOW = 2
-
-function stripSalvage(diagnosis: RuntimeTurnError['diagnosis']): SessionDeathDiagnosisLite {
-  return {
-    reason: diagnosis.reason,
-    ...(diagnosis.sessionId ? { sessionId: diagnosis.sessionId } : {}),
-    ...(diagnosis.sessionStatus ? { sessionStatus: diagnosis.sessionStatus } : {}),
-    ...(diagnosis.completionBytes !== undefined ? { completionBytes: diagnosis.completionBytes } : {}),
-    ...(diagnosis.outputTruncated !== undefined ? { outputTruncated: diagnosis.outputTruncated } : {}),
-    ...(diagnosis.oversizedOutput !== undefined ? { oversizedOutput: diagnosis.oversizedOutput } : {}),
-    ...(diagnosis.lastToolCall ? { lastToolCall: diagnosis.lastToolCall } : {}),
-    ...(diagnosis.detail ? { detail: diagnosis.detail } : {}),
-  }
-}
-
-/**
- * Persist the truncated completion text the adapter salvaged from the dead
- * session as a task-linked asset. The blocked/retried task then carries the
- * partial deliverable instead of losing it — and the corrective prompt can
- * point the agent at it so work isn't regenerated from scratch.
- */
-async function salvageSessionDeathOutput(input: {
-  contentDir: string
-  taskId: string
-  agent: string
-  seq: number
-  salvagedText: string
-}): Promise<string | undefined> {
-  try {
-    const dir = join(input.contentDir, 'tasks', 'salvage')
-    const { mkdirSync } = await import('fs')
-    mkdirSync(dir, { recursive: true })
-    const filePath = join(dir, `${input.taskId}-d${input.seq}.md`)
-    writeFileSync(filePath, input.salvagedText, 'utf-8')
-    const result = await hooks().invoke<{ assetId: string; version: number; changed: boolean }>('assets.saveFromSource', {
-      filePath,
-      taskId: input.taskId,
-      agent: input.agent,
-      type: 'text',
-      description: 'Partial output salvaged from a dead runtime session (truncated by the provider)',
-      tags: ['salvaged-output'],
-      tool: 'session-forensics',
-    })
-    return result?.assetId
-  } catch (err) {
-    log.warn('Failed to salvage session-death output as asset', err, { taskId: input.taskId })
-    return undefined
-  }
-}
-
-/**
- * The session-death recovery ladder. Diagnosed deaths are deterministic —
- * blind retries reproduce them (the originating incident proved it twice) —
- * so each rung changes the approach instead of waiting out a cooldown:
- *
- *   death 1 → salvage partial output → IMMEDIATE corrective re-dispatch
- *             (prompt explains why the attempt died + artifact-first steps)
- *   death 2 → salvage → decomposition dispatch (do NOT do the work; split
- *             into chained single-deliverable subtasks)   [regular only]
- *   death 3 → block with the full diagnosis and actionable next steps
- *
- * Workflow steps skip decomposition (corrective → block): step structure is
- * owned by the workflow engine.
- */
-async function handleSessionDeath(input: {
-  contentDir: string
-  port: number
-  state: DispatchState
-  task: DispatchTask
-  targetAgent: string
-  err: RuntimeTurnError
-  dispatchKind: 'regular' | 'workflow'
-  snapshotColumn: keyof DispatchColumns | null
-}): Promise<void> {
-  const diagnosis = input.err.diagnosis
-  if (!input.state.failedDispatches) input.state.failedDispatches = {}
-  const prev = input.state.failedDispatches[input.task.id]?.sessionDeath
-  const deaths = (prev?.deaths ?? 0) + 1
-  const seq = currentSeq(input.task.id)
-
-  const salvagedAssetId = diagnosis.salvagedText
-    ? await salvageSessionDeathOutput({
-        contentDir: input.contentDir,
-        taskId: input.task.id,
-        agent: input.targetAgent,
-        seq,
-        salvagedText: diagnosis.salvagedText,
-      })
-    : undefined
-  const salvagedAssetIds = [...(prev?.salvagedAssetIds ?? []), ...(salvagedAssetId ? [salvagedAssetId] : [])]
-
-  const sizeLabel = diagnosis.completionBytes !== undefined
-    ? `${Math.round(diagnosis.completionBytes / 1024)}KB${diagnosis.outputTruncated ? ' (truncated)' : ''}`
-    : 'unknown size'
-  const auditPayload = {
-    id: input.task.id,
-    title: input.task.title,
-    kind: input.dispatchKind,
-    deaths,
-    ...stripSalvage(diagnosis),
-    ...(salvagedAssetId ? { salvagedAssetId } : {}),
-  }
-  appendAudit(input.contentDir, 'task.runtime_session_died', input.targetAgent, auditPayload)
-
-  const maxDeaths = input.dispatchKind === 'workflow' ? MAX_SESSION_DEATHS_WORKFLOW : MAX_SESSION_DEATHS_REGULAR
-  if (deaths >= maxDeaths) {
-    delete input.state.failedDispatches[input.task.id]
-    const salvageNote = salvagedAssetIds.length > 0
-      ? ` Salvaged partial output: asset(s) ${salvagedAssetIds.join(', ')}.`
-      : ''
-    const reason = [
-      `Runtime session died ${deaths} time(s) before completion.`,
-      diagnosis.detail ?? `Session ${diagnosis.sessionStatus ?? 'ended'} after a ${sizeLabel} completion.`,
-      diagnosis.lastToolCall ? `Last tool call: ${diagnosis.lastToolCall}.` : '',
-      salvageNote,
-      'Next step: split this task into smaller single-deliverable subtasks or reduce its scope, then re-dispatch.',
-    ].filter(Boolean).join(' ')
-    await blockStoredTask(input.task.id, reason)
-    await tryAddTaskLog(input.task.id, 'system', `Session died ${deaths} time(s); recovery ladder exhausted. Task blocked for review.`, { sessionDeath: auditPayload })
-    appendAudit(input.contentDir, 'task.runtime_failed_blocked', input.targetAgent, auditPayload)
-    return
-  }
-
-  const stage: SessionDeathState['stage'] = deaths === 1 ? 'corrective' : 'decomposition'
-  const existing = input.state.failedDispatches[input.task.id]
-  input.state.failedDispatches[input.task.id] = {
-    lastAttempt: Date.now(),
-    count: existing?.count ?? 0, // session deaths don't burn generic retries
-    kind: 'structural',
-    sessionDeath: { stage, deaths, lastDiagnosis: stripSalvage(diagnosis), salvagedAssetIds },
-  }
-
-  await tryAddTaskLog(
-    input.task.id,
-    'system',
-    `Runtime session died (${sizeLabel} completion, ${diagnosis.sessionStatus ?? diagnosis.reason}). ${stage === 'corrective' ? 'Re-dispatching with corrective output-discipline guidance.' : 'Dispatching decomposition: the agent will split this task into single-deliverable subtasks.'}${salvagedAssetId ? ` Partial output salvaged as asset ${salvagedAssetId}.` : ''}`,
-    { sessionDeath: auditPayload },
-  )
-
-  if (input.dispatchKind === 'workflow') {
-    // Step re-dispatch happens through the normal in-progress workflow scan
-    // (markers were already cleared); the corrective prompt is injected from
-    // the persisted sessionDeath state. The task stays inProgress.
-    appendAudit(input.contentDir, 'task.corrective_redispatch', input.targetAgent, auditPayload)
-    return
-  }
-
-  if (input.snapshotColumn === 'inProgress') {
-    try {
-      await moveStoredTask(input.task.id, 'todo', 'inProgress')
-    } catch (err) {
-      log.warn('Failed to return task to todo after session death', err, { id: input.task.id })
-    }
-  }
-
-  appendAudit(
-    input.contentDir,
-    stage === 'corrective' ? 'task.corrective_redispatch' : 'task.decomposition_dispatched',
-    input.targetAgent,
-    auditPayload,
-  )
-
-  // Immediate ladder re-dispatch — fired after the state lock releases
-  // (dispatchSingleTask takes the same lock; calling it inline would
-  // deadlock). We know exactly what went wrong and how the next attempt
-  // differs; parking the task for the next 5-minute cycle buys nothing.
-  // If the timer is lost (process restart) or the dispatch is gate-deferred,
-  // the next cycle picks the task up from its persisted sessionDeath state.
-  pendingLadderRedispatches += 1
-  const timer = setTimeout(() => {
-    dispatchSingleTask(input.task.id, input.contentDir, input.port, 'recovery')
-      .catch((err) => {
-        log.error('Session-death ladder re-dispatch failed', err, { id: input.task.id, stage })
-      })
-      .finally(() => {
-        pendingLadderRedispatches -= 1
-      })
-  }, SESSION_DEATH_REDISPATCH_DELAY_MS)
-  timer.unref?.()
-}
-
-async function reconcileRejectedDispatch(input: {
-  contentDir: string
-  port: number
-  state: DispatchState
-  dispatchedSet: Set<string> | null
-  task: DispatchTask
-  targetAgent: string
-  err: unknown
-  initialLogCount: number
-  logPrefix: string
-  dispatchKind: 'regular' | 'workflow'
-}): Promise<void> {
-  const snapshot = findDispatchTaskSnapshot(input.task.id)
-  removeDispatchMarkersForTask(input.state, input.dispatchedSet, input.task.id)
-
-  if (snapshot && taskAlreadyLeftActiveWork(snapshot.column)) {
-    delete input.state.failedDispatches?.[input.task.id]
-    appendAudit(input.contentDir, 'task.dispatch_failure_ignored', input.targetAgent, {
-      id: input.task.id,
-      title: input.task.title,
-      column: snapshot.column,
-      error: formatSanitizedRuntimeFailure(input.err),
-    })
-    return
-  }
-
-  // Diagnosed session deaths take the recovery ladder — never the generic
-  // block-or-cooldown paths below.
-  if (input.err instanceof RuntimeTurnError) {
-    await handleSessionDeath({
-      contentDir: input.contentDir,
-      port: input.port,
-      state: input.state,
-      task: input.task,
-      targetAgent: input.targetAgent,
-      err: input.err,
-      dispatchKind: input.dispatchKind,
-      snapshotColumn: snapshot?.column ?? null,
-    })
-    return
-  }
-
-  const summary = formatSanitizedRuntimeFailure(input.err)
-  if (snapshot?.column === 'inProgress' && shouldBlockAfterDispatchFailure(snapshot, input.initialLogCount)) {
-    delete input.state.failedDispatches?.[input.task.id]
-    await blockStoredTask(input.task.id, `Agent run ended before reporting completion. ${summary}.`)
-    await tryAddTaskLog(input.task.id, 'system', `Agent run ended before task completion: ${summary}. Task moved to blocked for review.`)
-    appendAudit(input.contentDir, 'task.runtime_failed_blocked', input.targetAgent, {
-      id: input.task.id,
-      title: input.task.title,
-      kind: input.dispatchKind,
-      error: summary,
-    })
-    return
-  }
-
-  if (!input.state.failedDispatches) input.state.failedDispatches = {}
-  const prev = getFailureRecord(input.state.failedDispatches[input.task.id])
-  const kind = classifyDispatchError(input.err)
-  const detail = classifyDispatchFailureDetail(input.err)
-  const attempt = (prev?.count || 0) + 1
-  // Preserve any in-flight recovery-ladder state: a transport failure
-  // between ladder rungs (e.g. the corrective re-dispatch fired while the
-  // gateway was still rebooting after the death) must not wipe the stage —
-  // otherwise the ladder restarts from scratch and the corrective prompt
-  // is lost. Found live during the rig ladder smoke.
-  input.state.failedDispatches[input.task.id] = {
-    lastAttempt: Date.now(),
-    count: attempt,
-    kind,
-    ...(prev?.sessionDeath ? { sessionDeath: prev.sessionDeath } : {}),
-  }
-
-  if (snapshot?.column === 'inProgress') {
-    try {
-      await moveStoredTask(input.task.id, 'todo', 'inProgress')
-    } catch (err) {
-      log.warn('Failed to return task to todo after dispatch rejection', err, { id: input.task.id })
-    }
-  }
-
-  await tryAddTaskLog(
-    input.task.id,
-    'system',
-    `${input.logPrefix} (attempt ${attempt}, ${kind}) → ${input.targetAgent}: ${detail.summary}. Returned to Todo for retry when cooldown expires.`,
-    { dispatchFailure: detail },
-  )
-  appendAudit(input.contentDir, 'task.dispatch_failed', input.targetAgent, {
-    id: input.task.id,
-    title: input.task.title,
-    error: detail.summary,
-    attempt,
-    kind,
-    ...detail,
-  })
-}
-
-// ─── In-flight turn registry (concurrent dispatch) ─────────────────────────
-//
-// Sends are fired and tracked, never awaited inside the dispatch cycle: one
-// 10-minute turn must not stall dispatching to every other agent (the
-// serial-await loop was why parallel task fan-out made no independent
-// progress). Reconciliation happens in per-turn settle handlers that take
-// the state lock themselves. The registry is advisory — caps + settle
-// bookkeeping — never a source of truth for task state; restart safety is
-// unchanged (in-flight promises die with the process, restart-recovery
-// handles orphaned inProgress tasks via heartbeats).
-
-const inFlightTurns = new Map<string, InFlightTurn>()
-
-export function getInFlightTurnCount(agentId?: string): number {
-  if (!agentId) return inFlightTurns.size
-  let count = 0
-  for (const turn of inFlightTurns.values()) {
-    if (turn.agentId === agentId) count += 1
-  }
-  return count
-}
-
-/** Why a dispatch can't fire right now, or null when a slot is free. */
-function concurrencyGate(
-  agentId: string,
-  settings: { dispatch: { maxConcurrentTurns: number; maxTurnsPerAgent: number } },
-  // Slots reserved by collected-but-not-yet-fired turns in the current
-  // two-phase cycle — invisible to the in-flight registry until phase 2.
-  reserved?: { total: number; forAgent: number },
-): ConcurrencyGate {
-  if (inFlightTurns.size + (reserved?.total ?? 0) >= settings.dispatch.maxConcurrentTurns) return 'concurrency_cap'
-  if (getInFlightTurnCount(agentId) + (reserved?.forAgent ?? 0) >= settings.dispatch.maxTurnsPerAgent) return 'agent_busy'
-  return null
-}
-
-// Ladder re-dispatches scheduled via the post-lock timer but not yet fired —
-// without counting these, awaitDispatchIdle() would report idle during the
-// 50ms window between a dying turn settling and its recovery turn firing.
-let pendingLadderRedispatches = 0
-
-/**
- * Wait until every tracked turn (including its settle-time reconciliation)
- * and every scheduled ladder re-dispatch has finished. Deterministic
- * synchronization point for tests and shutdown.
- */
-export async function awaitDispatchIdle(): Promise<void> {
-  while (inFlightTurns.size > 0 || pendingLadderRedispatches > 0) {
-    if (inFlightTurns.size > 0) {
-      await Promise.allSettled([...inFlightTurns.values()].map((turn) => turn.settled))
-    } else {
-      await new Promise((resolve) => setTimeout(resolve, 10))
-    }
-  }
-}
-
-/**
- * Fire a dispatch send and reconcile on settle. Returns immediately. The
- * settle handlers re-load state under the lock — the firing cycle's state
- * object must not be touched after the cycle's own save.
- */
-function fireDispatchTurn(opts: {
-  marker: string
-  task: DispatchTask
-  targetAgent: string
-  threadId: string
-  message: string
-  contentDir: string
-  port: number
-  initialLogCount: number
-  logPrefix: string
-  dispatchKind: 'regular' | 'workflow'
-  /** True when this is a recovery-ladder re-dispatch (routes to the
-   *  'recovery' origin policy regardless of the task's shape). */
-  isRecovery?: boolean
-  onSettled?: (outcome: 'ok' | 'error', err?: unknown) => void
-}): void {
-  // The registry entry is released only AFTER settle reconciliation
-  // completes — awaitDispatchIdle() must mean "all bookkeeping done", and a
-  // turn's slot shouldn't free until its outcome has been recorded.
-  const settled = resolveDispatchRouting(opts.task, opts.isRecovery ?? false)
-    .then(async (routing) => {
-      if (routing.model || routing.thinking) {
-        appendAudit(opts.contentDir, 'task.routed', opts.targetAgent, {
-          id: opts.task.id,
-          ...(routing.model ? { model: routing.model } : {}),
-          ...(routing.thinking ? { thinking: routing.thinking } : {}),
-        })
-      }
-      const result = await sendDispatchMessage(opts.targetAgent, opts.message, opts.threadId, routing)
-      // Free the live-run slot FIRST — the settle reconciliation below (and
-      // any ladder re-dispatch it schedules) must be able to claim anew.
-      try {
-        settleRun(opts.threadId, 'ok')
-      } catch (err) {
-        log.error('Failed to settle run in ledger', err, { threadId: opts.threadId })
-      }
-      // Attribute the turn's token/dollar cost (run_id == threadId). The
-      // resolved routing model is what actually ran — price against it.
-      await recordTurnCost(opts.threadId, opts.task.id, opts.targetAgent, result, routing.model)
-      let completedDecomposition = false
-      await withStateLock(() => {
-        const state = loadDispatchState(opts.contentDir)
-        const record = state.failedDispatches?.[opts.task.id]
-        if (record) {
-          completedDecomposition = record.sessionDeath?.stage === 'decomposition'
-          delete state.failedDispatches![opts.task.id]
-          saveDispatchState(opts.contentDir, state)
-        }
-      })
-      // A successful DECOMPOSITION turn ends with the agent creating the
-      // subtask chain and STOPPING (no tasks_complete, by design) — which
-      // leaves the parent inProgress, where continuation skips it when the
-      // chain finishes (found live on the rig: the parent idled until
-      // watchdog recovery). Park it in todo; the dependsOn gate the agent
-      // set holds it there until the chain completes, then continuation
-      // re-dispatches it for final assembly.
-      if (completedDecomposition && findDispatchTaskSnapshot(opts.task.id)?.column === 'inProgress') {
-        try {
-          await moveStoredTask(opts.task.id, 'todo', 'inProgress')
-          await tryAddTaskLog(opts.task.id, 'system', 'Decomposition complete — parked in Todo until the subtask chain finishes (dependency gate), then re-dispatched for final assembly.')
-        } catch (err) {
-          log.warn('Failed to park decomposed parent in todo', err, { id: opts.task.id })
-        }
-      }
-      opts.onSettled?.('ok')
-    })
-    .catch(async (err) => {
-      log.error(`${opts.logPrefix}: turn failed for "${opts.task.title}" → ${opts.targetAgent}`, err)
-      // Free the slot before reconciliation — the recovery ladder's
-      // re-dispatch claims a fresh run and must not be suppressed by the
-      // dead one. Session deaths are 'lost'; other failures 'settled'.
-      try {
-        if (err instanceof RuntimeTurnError) loseRun(opts.threadId, 'session-death')
-        else settleRun(opts.threadId, `failed: ${formatDispatchError(err).slice(0, 120)}`)
-      } catch (ledgerErr) {
-        log.error('Failed to settle failed run in ledger', ledgerErr, { threadId: opts.threadId })
-      }
-      try {
-        await withStateLock(async () => {
-          const state = loadDispatchState(opts.contentDir)
-          await reconcileRejectedDispatch({
-            contentDir: opts.contentDir,
-            port: opts.port,
-            state,
-            dispatchedSet: null,
-            task: opts.task,
-            targetAgent: opts.targetAgent,
-            err,
-            initialLogCount: opts.initialLogCount,
-            logPrefix: opts.logPrefix,
-            dispatchKind: opts.dispatchKind,
-          })
-          saveDispatchState(opts.contentDir, state)
-        })
-      } catch (reconcileErr) {
-        log.error('Dispatch settle reconciliation failed', reconcileErr, { id: opts.task.id })
-      }
-      opts.onSettled?.('error', err)
-    })
-    .finally(() => {
-      inFlightTurns.delete(opts.marker)
-    })
-
-  inFlightTurns.set(opts.marker, {
-    agentId: opts.targetAgent,
-    taskId: opts.task.id,
-    threadId: opts.threadId,
-    startedAt: Date.now(),
-    settled,
-  })
-}
-
-/** Shared dispatched[] cap — both dispatch paths must honor the setting. */
-export async function dispatchTasks(contentDir: string, port: number): Promise<void> {
-  // If a previous dispatch is stuck (>3min), force-release the mutex
-  if (dispatching && dispatchStartedAt > 0 && (Date.now() - dispatchStartedAt) > DISPATCH_TIMEOUT_MS) {
-    log.warn('Dispatch mutex stuck — force-releasing after timeout', { stuckSinceMs: Date.now() - dispatchStartedAt })
-    dispatching = false
-  }
-  if (dispatching) return
-  dispatching = true
-  dispatchStartedAt = Date.now()
-
-  const settings = getSettings()
-
-  try {
-    // Acquire state lock for the entire cycle to prevent races with dispatchSingleTask
-    await withStateLock(async () => {
-    const state = loadDispatchState(contentDir)
-    const runtime = getAppServices().runtime
-    const runtimeAgents = await runtime.agents.list()
-    const runtimeAgentIds = new Set(runtimeAgents.map((agent) => agent.id))
-    const runtimeRoster: DispatchRosterAgent[] = runtimeAgents.map((agent) => ({
-      id: agent.id,
-      ...(agent.role ? { role: agent.role } : {}),
-    }))
-    const mainAgentId = await getRuntimeMainAgentId(runtime)
-    // Reconcile dispatch state with taskboard reality
-    const columns = await readDispatchColumns()
-    const todoTasks = columns.todo
-    const activeIds = new Set(columns.inProgress.map(t => t.id))
-    const completedTaskIds = new Set([
-      ...columns.done.map(t => t.id),
-      ...columns.archived.map(t => t.id),
-    ])
-    const knownTaskIds = new Set(
-      Object.values(columns).flatMap((column) => column.map((t) => t.id)),
-    )
-    for (const task of [
-      ...columns.done,
-      ...columns.archived,
-      ...columns.blocked,
-      ...columns.review,
-    ]) {
-      delete state.failedDispatches?.[task.id]
-    }
-    state.dispatched = state.dispatched.filter(id => activeIds.has(getDispatchMarkerTaskId(id)))
-    // Rebuild dispatchedSet AFTER reconciliation so tasks moved back to todo are eligible
-    const dispatchedSet = new Set(state.dispatched)
-    for (const task of columns.inProgress) {
-      if (!dispatchedSet.has(task.id)) {
-        dispatchedSet.add(task.id)
-        state.dispatched.push(task.id)
-      }
-    }
-
-    // ─── Dispatch in-progress workflow tasks that changed step agents ──
-    // After gate approval, the workflow advances to a new agent's step but
-    // the task stays in inProgress. We need to dispatch the new agent.
-    for (const task of columns.inProgress) {
-      const wfTask = task as typeof task & { workflowId?: string }
-      if (!wfTask.workflowId) continue
-
-      // Check if the current workflow step agent has been dispatched
-      const activeAgents = await hooks().invoke<Array<{ agent: string; stepId: string; effectiveTaskId?: string }>>('workflows.getActiveAgents', { taskId: task.id }) ?? []
-      const needsDispatch = activeAgents.some(({ stepId }) => !dispatchedSet.has(`${task.id}:${stepId}`))
-      if (!needsDispatch) continue
-
-      try {
-        await dispatchWorkflowTask(
-          { ...wfTask, workflowId: wfTask.workflowId },
-          contentDir, port, dispatchedSet, state,
-          async () => {}, // already in progress, no column move needed
-          addTaskLog,
-          mainAgentId,
-        )
-      } catch (err) {
-        log.error(`Failed to re-dispatch workflow task "${task.title}"`, err)
-      }
-    }
-
-    if (todoTasks.length === 0) {
-      saveDispatchState(contentDir, state)
-      return
-    }
-
-    // Two-phase dispatch: phase 1 collects intents (move + mint + message
-    // build, seqs minted in-memory via deferSave), then ONE state save
-    // persists every minted seq, then phase 2 fires the turns. This keeps
-    // persist-before-send (a seq is always durable before its turn is sent)
-    // while collapsing the old N+1 full-file state writes per cycle to 1.
-    // A mid-loop failure loses only unminted work: collected intents have
-    // not fired yet, and unfired seqs re-mint identically next cycle.
-    const pendingTurns: Array<Parameters<typeof fireDispatchTurn>[0]> = []
-    const pendingByAgent = new Map<string, number>()
-    // Per-cycle memo for budget spend reads (global totals are identical for
-    // every task in this cycle; costs only land on settle, after the loop).
-    const budgetSpendCache = new Map<string, number>()
-
-    for (const task of todoTasks) {
-      if (dispatchedSet.has(task.id)) continue
-
-      const eligibility = isTaskDispatchEligible(task, {
-        nowMs: Date.now(),
-        runtimeAgentIds,
-        completedTaskIds,
-        knownTaskIds,
-      })
-      if (!eligibility.eligible) continue
-      if (eligibility.danglingDependency) {
-        log.warn('Task dependency target no longer exists; treating as satisfied', { id: task.id, dependsOn: eligibility.danglingDependency })
-        await tryAddTaskLog(task.id, 'system', `Dependency task ${eligibility.danglingDependency} no longer exists (likely archived/removed). Treating the dependency as satisfied and dispatching.`)
-      }
-
-      // Check failure history with max retries. Session-death records are
-      // ladder-managed (corrective/decomposition with their own caps) and
-      // bypass the generic cooldown — waiting doesn't fix a deterministic
-      // failure; the changed approach does.
-      const failure = getFailureRecord(state.failedDispatches?.[task.id])
-      if (failure && !failure.sessionDeath) {
-        if (failure.count >= settings.dispatch.maxRetries) {
-          // Escalate to blocked
-          try {
-            await blockStoredTask(task.id, `Dispatch failed ${failure.count} times - agent may be unavailable`)
-            await addTaskLog(task.id, 'system', `Dispatch exhausted: ${failure.count} failed attempts. Task moved to blocked.`)
-            appendAudit(contentDir, 'task.dispatch_exhausted', 'system', { id: task.id, title: task.title, count: failure.count })
-            log.warn('Task blocked after max dispatch retries', { id: task.id, count: failure.count })
-          } catch (err) {
-            log.error('Failed to block exhausted task', err)
-          }
-          continue
-        }
-        if (Date.now() - failure.lastAttempt < cooldownForFailure(failure, settings)) continue
-      }
-
-      // Workflow-aware dispatch path
-      const taskWithWorkflow = task as typeof task & { workflowId?: string }
-      if (taskWithWorkflow.workflowId) {
-        try {
-          await dispatchWorkflowTask({ ...taskWithWorkflow, workflowId: taskWithWorkflow.workflowId }, contentDir, port, dispatchedSet, state, moveTaskToInProgress, addTaskLog, mainAgentId)
-        } catch (err) {
-          log.error(`Failed to dispatch workflow task "${task.title}"`, err)
-        }
-        continue
-      }
-
-      const targetAgent = task.agent ?? mainAgentId
-
-      // Bounded concurrency: a capped task is simply ineligible this cycle
-      // (no failure recorded) — a later cycle picks it up when a slot frees.
-      // Collected-but-unfired turns reserve their slots via pendingByAgent.
-      const gate = concurrencyGate(targetAgent, settings, {
-        total: pendingTurns.length,
-        forAgent: pendingByAgent.get(targetAgent) ?? 0,
-      })
-      if (gate) {
-        log.debug('Dispatch deferred by concurrency gate', { id: task.id, agent: targetAgent, gate })
-        continue
-      }
-
-      // Spend ceiling: defer (leave in todo) when a budget cap is hit. Runs
-      // before the claim so we don't reserve a run we won't fire. The
-      // per-cycle cache collapses the redundant global spend reads.
-      if (await deferForBudget(targetAgent, contentDir, budgetSpendCache)) {
-        log.debug('Dispatch deferred by budget gate', { id: task.id, agent: targetAgent })
-        continue
-      }
-
-      // Per-task guard: one bad task (store hiccup, hook failure) must not
-      // abort the cycle — collected intents and other tasks still dispatch.
-      let claim: ClaimNextRunResult | null = null
-      try {
-        // Claim FIRST — the ledger run row is the lock; everything after it
-        // is a side effect. A duplicate (overlapping cycle, racing kick,
-        // second process) fails here before any move or send.
-        claim = claimDispatchRun(task.id, targetAgent)
-        if (!claim.claimed) {
-          auditDispatchSuppressed(contentDir, task, targetAgent, claim.liveRunId, 'cycle')
-          continue
-        }
-
-        const lessonBlock = await buildDispatchLessonBlock({
-          contentDir,
-          taskId: task.id,
-          title: task.title,
-          agentId: targetAgent,
-          query: buildTaskLessonQuery(task),
-        })
-        const assetsBlock = await buildDispatchAssetBlock(task.id)
-        const recovery = failure?.sessionDeath
-        const message = recovery?.stage === 'decomposition'
-          ? buildDecompositionMessage(task, targetAgent, recovery)
-          : buildDispatchMessage(task, targetAgent, contentDir, mainAgentId, lessonBlock, {}, recovery, runtimeRoster, assetsBlock)
-        const initialLogCount = task.log?.length ?? 0
-
-        // Move to inProgress BEFORE sending message to eliminate race condition
-        // where fast agents complete before dispatch moves the task
-        await moveTaskToInProgress(task.id, targetAgent)
-
-        const threadId = claim.runId
-        dispatchedSet.add(task.id)
-        // The internal todo→inProgress move is folded into task.dispatched —
-        // one audit row per dispatch, carrying the transition.
-        appendAudit(contentDir, 'task.dispatched', targetAgent, { id: task.id, title: task.title, threadId, from: 'todo', to: 'inProgress' })
-        log.info('Task dispatched', { id: task.id, title: task.title, agent: targetAgent, threadId })
-        pendingTurns.push({
-          marker: task.id,
-          task,
-          targetAgent,
-          threadId,
-          message,
-          contentDir,
-          port,
-          initialLogCount,
-          logPrefix: 'Dispatch failed',
-          dispatchKind: 'regular',
-          // A task carrying a sessionDeath record is a recovery re-dispatch
-          // even on the main cycle (e.g. the ladder's immediate re-dispatch
-          // was lost to a restart or budget-deferred). Route it to the
-          // 'recovery' origin just like dispatchSingleTask(...,'recovery').
-          isRecovery: !!recovery,
-        })
-        pendingByAgent.set(targetAgent, (pendingByAgent.get(targetAgent) ?? 0) + 1)
-      } catch (err) {
-        log.error(`Failed to prepare dispatch for task "${task.title}"`, err, { id: task.id })
-        // A claim whose turn will never fire must be released, or the task
-        // is locked until the watchdog/boot sweep notices.
-        if (claim?.claimed) loseRun(claim.runId, 'dispatch-prep-failed')
-      }
-    }
-
-    state.lastRun = Date.now()
-    state.dispatched = [...dispatchedSet]
-    trimDispatched(state, settings.dispatch.maxDispatched)
-    // Persist-before-send: every claimed run above is already durable in the
-    // ledger; this write persists the advisory dispatch bookkeeping. If it
-    // throws, the collected turns will never fire — release their claims or
-    // the tasks stay locked until the watchdog/boot sweep notices.
-    try {
-      saveDispatchState(contentDir, state)
-    } catch (err) {
-      log.error('Dispatch state save failed — releasing collected run claims', err)
-      for (const turn of pendingTurns) {
-        try {
-          loseRun(turn.threadId, 'dispatch-save-failed')
-        } catch (releaseErr) {
-          log.error('Failed to release collected claim', releaseErr, { threadId: turn.threadId })
-        }
-      }
-      return
-    }
-
-    for (const turn of pendingTurns) fireDispatchTurn(turn)
-    }) // end withStateLock
-  } finally {
-    dispatching = false
-  }
-}
-
-/**
- * Immediately dispatch a single task by ID, bypassing the 5-minute cycle.
- * Used for kick (explicit) and auto-kick (subtask with parentId).
- */
-/**
- * Remove a task's dispatch markers so an immediate re-dispatch (e.g.
- * dependency continuation) isn't skipped as "already dispatched" before the
- * next cycle reconciles markers against the board.
- */
-export async function dispatchSingleTask(
-  taskId: string,
-  contentDir: string,
-  port: number,
-  source: 'kick' | 'subtask' | 'continuation' | 'recovery' = 'kick',
-  continuation: DispatchContinuationContext = {},
-): Promise<void> {
-  const settings = getSettings()
-
-  await withStateLock(async () => {
-    const columns = await readDispatchColumns()
-    const task = columns.todo.find(t => t.id === taskId)
-    if (!task) {
-      log.debug('dispatchSingleTask: task not in todo, skipping', { taskId })
-      return
-    }
-
-    const runtime = getAppServices().runtime
-    const runtimeAgents = await runtime.agents.list()
-    const runtimeAgentIds = new Set(runtimeAgents.map((agent) => agent.id))
-    const runtimeRoster: DispatchRosterAgent[] = runtimeAgents.map((agent) => ({
-      id: agent.id,
-      ...(agent.role ? { role: agent.role } : {}),
-    }))
-    const mainAgentId = await getRuntimeMainAgentId(runtime)
-    const completedTaskIds = new Set([
-      ...columns.done.map(t => t.id),
-      ...columns.archived.map(t => t.id),
-    ])
-    const knownTaskIds = new Set(
-      Object.values(columns).flatMap((column) => column.map((t) => t.id)),
-    )
-    const eligibilityTask = source === 'kick' ? { ...task, availableAt: undefined } : task
-    const eligibility = isTaskDispatchEligible(eligibilityTask, {
-      nowMs: Date.now(),
-      runtimeAgentIds,
-      completedTaskIds,
-      knownTaskIds,
-    })
-    if (!eligibility.eligible) {
-      log.debug('dispatchSingleTask: task not dispatch eligible', { taskId, reason: eligibility.reason })
-      return
-    }
-    if (eligibility.danglingDependency) {
-      log.warn('Task dependency target no longer exists; treating as satisfied', { id: taskId, dependsOn: eligibility.danglingDependency })
-      await tryAddTaskLog(taskId, 'system', `Dependency task ${eligibility.danglingDependency} no longer exists (likely archived/removed). Treating the dependency as satisfied and dispatching.`)
-    }
-
-    const state = loadDispatchState(contentDir)
-    if (state.dispatched.includes(taskId)) {
-      log.debug('dispatchSingleTask: already dispatched', { taskId })
-      return
-    }
-
-    // Check failure history (session-death records are ladder-managed and
-    // exempt from the generic cooldown/retry caps — see dispatchTasks).
-    const failure = getFailureRecord(state.failedDispatches?.[taskId])
-    if (failure && !failure.sessionDeath) {
-      if (failure.count >= settings.dispatch.maxRetries) {
-        log.warn('dispatchSingleTask: task exhausted retries', { taskId, count: failure.count })
-        return
-      }
-      if (Date.now() - failure.lastAttempt < cooldownForFailure(failure, settings)) {
-        log.debug('dispatchSingleTask: task in cooldown', { taskId, kind: failure.kind })
-        return
-      }
-    }
-
-    // Workflow-aware dispatch path
-    const taskWithWorkflow = task as typeof task & { workflowId?: string }
-    if (taskWithWorkflow.workflowId) {
-      const dispatchedSet = new Set(state.dispatched)
-      const wfStart = Date.now()
-      const wfAgent = task.agent ?? mainAgentId
-      try {
-        await dispatchWorkflowTask(
-          { ...taskWithWorkflow, workflowId: taskWithWorkflow.workflowId },
-          contentDir, port, dispatchedSet, state, moveTaskToInProgress, addTaskLog, mainAgentId,
-        )
-        state.dispatched = [...dispatchedSet]
-        saveDispatchState(contentDir, state)
-        appendAudit(contentDir, 'task.kicked', source, { id: taskId, title: task.title, workflow: true })
-        log.info('Single-task dispatch (workflow)', { id: taskId, title: task.title, source })
-        recordUsage({
-          kind: 'agent',
-          name: 'dispatch',
-          agent: wfAgent,
-          durationMs: Date.now() - wfStart,
-          status: 'ok',
-          meta: { taskId: task.id, title: task.title, workflow: true, source },
-        })
-      } catch (err) {
-        log.error(`dispatchSingleTask: workflow dispatch failed for "${task.title}"`, err)
-        recordUsage({
-          kind: 'agent',
-          name: 'dispatch',
-          agent: wfAgent,
-          durationMs: Date.now() - wfStart,
-          status: 'error',
-          meta: { taskId: task.id, title: task.title, workflow: true, source, error: formatDispatchError(err) },
-        })
-      }
-      return
-    }
-
-    // Regular task dispatch
-    const targetAgent = task.agent ?? mainAgentId
-
-    const gate = concurrencyGate(targetAgent, settings)
-    if (gate) {
-      if (source === 'recovery') {
-        // The "immediate" ladder re-dispatch degrades to next-cycle pickup —
-        // make that observable instead of a silent debug line.
-        log.warn('Ladder recovery dispatch deferred by concurrency gate; next cycle will retry', { taskId, agent: targetAgent, gate })
-        await tryAddTaskLog(taskId, 'system', `Recovery re-dispatch deferred (${gate}); the next dispatch cycle will retry.`)
-      } else {
-        log.debug('dispatchSingleTask: deferred by concurrency gate', { taskId, agent: targetAgent, gate })
-      }
-      return
-    }
-
-    const lessonBlock = await buildDispatchLessonBlock({
-      contentDir,
-      taskId: task.id,
-      title: task.title,
-      agentId: targetAgent,
-      query: buildTaskLessonQuery(task),
-    })
-    const assetsBlock = await buildDispatchAssetBlock(task.id)
-    const recovery = failure?.sessionDeath
-    const message = recovery?.stage === 'decomposition'
-      ? buildDecompositionMessage(task, targetAgent, recovery)
-      : buildDispatchMessage(task, targetAgent, contentDir, mainAgentId, lessonBlock, continuation, recovery, runtimeRoster, assetsBlock)
-    const dispatchStart = Date.now()
-    const initialLogCount = task.log?.length ?? 0
-
-    // Spend ceiling — defer (leave in todo) when a budget cap is hit.
-    if (await deferForBudget(targetAgent, contentDir)) {
-      log.debug('Single-task dispatch deferred by budget gate', { id: task.id, agent: targetAgent, source })
-      return
-    }
-
-    // Claim first — the ledger row is the lock (a kick racing the cycle or
-    // another process fails here, before any side effect).
-    const claim = claimDispatchRun(task.id, targetAgent)
-    if (!claim.claimed) {
-      auditDispatchSuppressed(contentDir, task, targetAgent, claim.liveRunId, 'single')
-      return
-    }
-
-    const threadId = claim.runId
-    try {
-      // Move to inProgress BEFORE sending message to eliminate race condition
-      await moveTaskToInProgress(task.id, targetAgent)
-
-      state.dispatched.push(task.id)
-      trimDispatched(state, settings.dispatch.maxDispatched)
-      saveDispatchState(contentDir, state)
-    } catch (err) {
-      // A claim whose turn will never fire must be released, or the task is
-      // locked until the watchdog/boot sweep notices (mirrors the cycle's
-      // per-task guard).
-      try {
-        loseRun(threadId, 'dispatch-prep-failed')
-      } catch (releaseErr) {
-        log.error('Failed to release claim after prep failure', releaseErr, { threadId })
-      }
-      throw err
-    }
-
-    // Internal move folded into task.dispatched — one audit row per dispatch.
-    appendAudit(contentDir, 'task.dispatched', targetAgent, { id: task.id, title: task.title, threadId, from: 'todo', to: 'inProgress' })
-    if (source !== 'continuation') {
-      appendAudit(contentDir, 'task.kicked', source, { id: task.id, title: task.title })
-    }
-    log.info('Single-task dispatch', { id: task.id, title: task.title, agent: targetAgent, source, threadId })
-    fireDispatchTurn({
-      marker: task.id,
-      task,
-      targetAgent,
-      threadId,
-      message,
-      contentDir,
-      port,
-      initialLogCount,
-      logPrefix: 'Immediate dispatch failed',
-      dispatchKind: 'regular',
-      isRecovery: source === 'recovery',
-      onSettled: (outcome, err) => {
-        recordUsage({
-          kind: 'agent',
-          name: 'dispatch',
-          agent: targetAgent,
-          durationMs: Date.now() - dispatchStart,
-          status: outcome,
-          meta: {
-            taskId: task.id,
-            title: task.title,
-            source,
-            ...(err ? { error: formatDispatchError(err) } : {}),
-          },
-        })
-      },
-    })
-  })
-}
-
-export function start(contentDir: string, port: number): void {
-  const settings = getSettings()
-  dispatchTimer = setInterval(() => {
-    dispatchTasks(contentDir, port).catch(err => {
-      log.error('Dispatch cycle failed', err)
-      appendAudit(contentDir, 'system.dispatch_error', 'system', { error: err instanceof Error ? err.message : String(err) })
-    })
-  }, settings.dispatch.intervalMs)
-  log.info('Dispatch started', { intervalMs: settings.dispatch.intervalMs })
-}
-
-export function stop(): void {
-  if (dispatchTimer) {
-    clearInterval(dispatchTimer)
-    dispatchTimer = null
-    log.info('Dispatch stopped')
-  }
-}
-
-/**
- * Dispatch a workflow-backed task. Creates an instance if needed,
- * then sends only the current step's instructions to the assigned agent.
- */
-async function dispatchWorkflowTask(
-  task: { id: string; title: string; description?: string; agent?: string; workflowId: string },
-  contentDir: string,
-  port: number,
-  dispatchedSet: Set<string>,
-  state: DispatchState,
-  moveTaskToInProgress: (id: string, agent: string) => Promise<void>,
-  addTaskLog: (id: string, author: string, message: string) => Promise<void>,
-  mainAgentId: string,
-): Promise<void> {
-  // Load or create workflow instance.
-  // Pass the task assignee so $assigned steps resolve to whoever owns the task at start time.
-  const instance = await hooks().invoke<Record<string, unknown>>('workflows.loadInstance', { taskId: task.id })
-  if (!instance) {
-    await hooks().invoke<Record<string, unknown>>('workflows.createInstance', { taskId: task.id, workflowId: task.workflowId, assignee: task.agent })
-    log.info('Created workflow instance', { taskId: task.id, workflowId: task.workflowId, resolvedAgent: task.agent })
-  } else if (!instance.resolvedAgent && task.agent) {
-    // Backfill resolvedAgent if instance was created before $assigned resolution was wired up
-    instance.resolvedAgent = task.agent
-    await hooks().invoke<void>('workflows.saveInstance', { instance })
-    log.info('Backfilled resolvedAgent on existing instance', { taskId: task.id, resolvedAgent: task.agent })
-  }
-
-  // Get agents for current step
-  const activeAgents = await hooks().invoke<Array<{ agent: string; stepId: string; effectiveTaskId?: string }>>('workflows.getActiveAgents', { taskId: task.id }) ?? []
-  if (activeAgents.length === 0) {
-    log.debug('No active agents for workflow step', { taskId: task.id })
-    return
-  }
-
-  // Move task to in_progress BEFORE dispatching to agents — same pattern as
-  // non-workflow tasks. Prevents the task from sitting in "todo" while the
-  // agent is already working on it.
-  if (!dispatchedSet.has(task.id)) {
-    const { columns: fresh } = readTaskboard() as unknown as { columns: Record<string, Array<{ id: string; agent?: string; title?: string; workflowId?: string }>> }
-    const stillInTodo = fresh.todo.some(t => t.id === task.id)
-    if (stillInTodo) {
-      const ownerAgent = task.agent || activeAgents[0]?.agent || mainAgentId
-      await moveTaskToInProgress(task.id, ownerAgent)
-      // No task.moved audit: the internal move is folded into the per-step
-      // task.dispatched rows emitted below.
-    }
-    dispatchedSet.add(task.id)
-  }
-
-  for (const { agent, stepId, effectiveTaskId } of activeAgents) {
-    // For nested workflows, use the child's taskId for step context resolution
-    const contextTaskId = effectiveTaskId || task.id
-    const stepContext = await hooks().invoke<Record<string, unknown>>('workflows.getCurrentStep', { taskId: contextTaskId, agentId: agent })
-    if (!stepContext || 'status' in stepContext && (stepContext.status === 'complete' || stepContext.status === 'pending_approval')) {
-      continue
-    }
-
-    const ctx = stepContext as {
-      stepId: string; label: string; type?: string; instructions?: string;
-      output_schema?: Record<string, unknown>; rejectionReason?: string;
-      previousOutput?: Record<string, unknown>; priorStepOutput?: Record<string, unknown>;
-      stepOutputs?: Record<string, Record<string, unknown>>; deny_tools?: string[]
-    }
-    const targetAgent = agent
-    const lessonBlock = await buildDispatchLessonBlock({
-      contentDir,
-      taskId: task.id,
-      title: task.title,
-      agentId: targetAgent,
-      query: buildTaskLessonQuery({
-        title: task.title,
-        description: task.description,
-        instructions: ctx.instructions,
-        context: ctx.label,
-      }),
-    })
-    // Pass contextTaskId so the step/complete API targets the right instance.
-    // A pending session-death recovery injects corrective guidance (workflow
-    // steps only get the corrective rung — the engine owns step structure).
-    const wfRecovery = getFailureRecord(state.failedDispatches?.[task.id])?.sessionDeath
-    // Assets are linked to the PARENT task id (decomposition guidance tells
-    // agents to save against the parent), so resolve by task.id, not contextTaskId.
-    const wfAssetsBlock = await buildDispatchAssetBlock(task.id)
-    const message = buildWorkflowDispatchMessage({ ...task, id: contextTaskId }, ctx, agent, lessonBlock, wfRecovery, wfAssetsBlock)
-    const initialLogCount = findDispatchTaskSnapshot(task.id)?.task.log?.length ?? 0
-
-    const gate = concurrencyGate(targetAgent, getSettings())
-    if (gate) {
-      log.debug('Workflow step dispatch deferred by concurrency gate', { taskId: task.id, stepId, agent: targetAgent, gate })
-      continue
-    }
-
-    // Spend ceiling — defer the step when a budget cap is hit.
-    if (await deferForBudget(targetAgent, contentDir)) {
-      log.debug('Workflow step dispatch deferred by budget gate', { taskId: task.id, stepId, agent: targetAgent })
-      continue
-    }
-
-    // Claim the step's live-run slot (exec key taskId:stepId — parallel
-    // step agents on one task are legitimate concurrent runs).
-    const claim = claimDispatchRun(task.id, targetAgent, stepId)
-    if (!claim.claimed) {
-      auditDispatchSuppressed(contentDir, task, targetAgent, claim.liveRunId, 'workflow')
-      continue
-    }
-
-    const threadId = claim.runId
-    dispatchedSet.add(`${task.id}:${stepId}`)
-    appendAudit(contentDir, 'task.dispatched', targetAgent, {
-      id: task.id,
-      title: task.title,
-      workflowId: task.workflowId,
-      stepId,
-      threadId,
-      from: 'todo',
-      to: 'inProgress',
-    })
-    log.info('Workflow step dispatched', { taskId: task.id, stepId, agent: targetAgent, threadId })
-    fireDispatchTurn({
-      marker: `${task.id}:${stepId}`,
-      task,
-      targetAgent,
-      threadId,
-      message,
-      contentDir,
-      port,
-      initialLogCount,
-      logPrefix: `Workflow dispatch failed for step "${stepId}"`,
-      dispatchKind: 'workflow',
-    })
-  }
-}
-
-/**
- * Build a dispatch message for a workflow step.
- * Contains ONLY the current step instructions — never future steps.
- *
- * Structure: identity frame → hard constraints → revision context → task → output → commands → stop.
- * Rules come BEFORE instructions (position primacy — LLMs weight early text more).
- */
-function buildWorkflowDispatchMessage(
-  task: { id: string; title: string; description?: string },
-  stepContext: {
-    stepId: string
-    label: string
-    type?: string
-    instructions?: string
-    output_schema?: Record<string, unknown>
-    rejectionReason?: string
-    previousOutput?: Record<string, unknown>
-    priorStepOutput?: Record<string, unknown>
-    stepOutputs?: Record<string, Record<string, unknown>>
-    deny_tools?: string[]
-  },
-  agentName: string,
-  lessonBlock = '',
-  recovery?: SessionDeathState,
-  assetsBlock = '',
-): string {
-  const lines: string[] = []
-  if (recovery?.stage === 'corrective') {
-    lines.push(buildCorrectiveSection(task.id, recovery).trimEnd())
-    lines.push('')
-  }
-
-  // ─── Identity Frame ─────────────────────────────────────────────────
-  lines.push('# WORKFLOW STEP ASSIGNMENT')
-  lines.push('')
-  lines.push('You are executing a single step in a managed workflow. You are NOT a general assistant right now — you are a workflow step executor.')
-  lines.push('')
-  lines.push(`**Task:** "${task.title}"`)
-  if (task.description) {
-    lines.push(`**Context:** ${task.description}`)
-  }
-  lines.push(`**Your step:** ${stepContext.label} (ID: ${stepContext.stepId})`)
-  lines.push(`**Your agent name:** ${agentName}`)
-  lines.push('')
-
-  // ─── Hard Constraints ───────────────────────────────────────────────
-  lines.push('## HARD CONSTRAINTS — violations are rejected server-side')
-  lines.push('')
-  lines.push('1. **SCOPE:** Do ONLY the work described in "YOUR TASK" below. Nothing more. If the task implies work for another step (e.g., generating images when your step is writing copy), STOP — that belongs to a different agent.')
-  lines.push('2. **OUTPUT:** Submit via bakin_exec_submit_step. Describing results in conversation does NOT complete the step. The workflow will not advance.')
-  lines.push('3. **SCHEMA:** Your output MUST match the JSON schema below. The server validates it. Missing fields = rejection. Extra fields = rejection. Wrong types = rejection.')
-  lines.push('4. **NO SIDE EFFECTS:** Do not create subtasks, dispatch other agents, move the task to Done, or post to any channel. The workflow engine handles ALL downstream handoffs — the next agent is already defined in the workflow and will be dispatched automatically when your step is approved. Creating a subtask would duplicate the workflow\'s job.')
-  lines.push('5. **ONE SUBMISSION:** Submit your output once via the API, then stop. Do not continue working after submission.')
-  if (stepContext.deny_tools?.length) {
-    lines.push(`6. **TOOL RESTRICTIONS:** Do NOT use: ${stepContext.deny_tools.join(', ')}. If this step requires those capabilities, BLOCK the task immediately.`)
-  }
-  lines.push('')
-  lines.push(...outputDisciplineSection(agentName, task.id, { subtasksAllowed: false }))
-  lines.push('')
-
-  // ─── Revision Context ───────────────────────────────────────────────
-  if (stepContext.rejectionReason) {
-    lines.push('## REVISION REQUIRED')
-    lines.push('')
-    if (stepContext.previousOutput) {
-      lines.push('**Previous output (rejected):**')
-      lines.push('```json')
-      lines.push(JSON.stringify(stepContext.previousOutput, null, 2))
-      lines.push('```')
-      lines.push('')
-    }
-    lines.push(`**What to fix:** ${stepContext.rejectionReason}`)
-    lines.push('')
-    lines.push('You MUST address this specific feedback. Do NOT resubmit unchanged output — the server detects near-duplicate resubmissions and rejects them.')
-    lines.push('')
-  }
-
-  // ─── Workflow Context (all prior step outputs) ─────────────────────
-  if (stepContext.stepOutputs && Object.keys(stepContext.stepOutputs).length > 0) {
-    lines.push('## WORKFLOW CONTEXT')
-    lines.push('')
-    lines.push('All completed step outputs from this workflow. Use as context for your work:')
-    lines.push('')
-    for (const [sid, output] of Object.entries(stepContext.stepOutputs)) {
-      const label = sid === '__parentContext' ? 'Parent Workflow (upstream handoff)' : `Step: ${sid}`
-      lines.push(`### ${label}`)
-      // Surface parent task metadata prominently for child workflows
-      if (sid === '__parentContext' && output && typeof output === 'object') {
-        const ctx = output as Record<string, unknown>
-        if (ctx._parentTaskTitle) {
-          lines.push(`**Parent Task:** ${ctx._parentTaskTitle}`)
-        }
-        if (ctx._parentTaskDescription) {
-          lines.push(`**Description:** ${ctx._parentTaskDescription}`)
-        }
-        // Show remaining output data (excluding internal metadata keys)
-        const rest = Object.fromEntries(Object.entries(ctx).filter(([k]) => !k.startsWith('_parent')))
-        if (Object.keys(rest).length > 0) {
-          lines.push('```json')
-          lines.push(JSON.stringify(rest, null, 2))
-          lines.push('```')
-        }
-      } else {
-        lines.push('```json')
-        lines.push(JSON.stringify(output, null, 2))
-        lines.push('```')
-      }
-      lines.push('')
-    }
-  } else if (stepContext.priorStepOutput) {
-    lines.push('## PRIOR STEP OUTPUT')
-    lines.push('')
-    lines.push('The previous step in this workflow produced the following output. Use this as context for your work:')
-    lines.push('```json')
-    lines.push(JSON.stringify(stepContext.priorStepOutput, null, 2))
-    lines.push('```')
-    lines.push('')
-  }
-
-  if (lessonBlock) {
-    lines.push(lessonBlock)
-    lines.push('')
-  }
-
-  if (assetsBlock) {
-    lines.push(assetsBlock.trim())
-    lines.push('')
-  }
-
-  // ─── Task Instructions ──────────────────────────────────────────────
-  lines.push('## YOUR TASK')
-  lines.push('')
-  if (stepContext.instructions) {
-    lines.push(stepContext.instructions)
-    lines.push('')
-  }
-
-  // ─── Required Output ────────────────────────────────────────────────
-  if (stepContext.output_schema) {
-    lines.push('## REQUIRED OUTPUT')
-    lines.push('')
-    lines.push('Your output must be valid JSON matching this schema:')
-    lines.push('```json')
-    lines.push(JSON.stringify(stepContext.output_schema, null, 2))
-    lines.push('```')
-    lines.push('')
-  }
-
-  // ─── Progress Logging ──────────────────────────────────────────────
-  lines.push('## PROGRESS LOGGING — MANDATORY')
-  lines.push('')
-  const { server: wfServer, mc: wfMc } = mcporterHelpers(agentName)
-
-  lines.push('You MUST log your progress throughout this workflow step. These updates appear in the live activity feed so humans can monitor your work in real-time.')
-  lines.push('')
-  lines.push('**When to log:**')
-  lines.push('- IMMEDIATELY when you start working (what you are about to do and your approach)')
-  lines.push('- After each significant action (reading files, generating content, making API calls, reviewing output)')
-  lines.push('- Share your reasoning ("The brief calls for warm tones, going with golden hour lighting")')
-  lines.push('- If anything unexpected happens or you are blocked')
-  lines.push('- When you complete and submit your output (summary of what you produced)')
-  lines.push('- If more than 2 minutes have passed since your last log, send a status update — even if just "Still working on X, currently Y"')
-  lines.push('')
-
-  // ─── Commands ───────────────────────────────────────────────────────
-  lines.push('## COMMANDS')
-  lines.push('')
-  lines.push(`Your Bakin MCP server is \`${wfServer}\`. Use mcporter for all interactions:`)
-  lines.push('')
-  lines.push('```bash')
-  lines.push(`# Submit your output (must match the schema above)`)
-  lines.push(`${wfMc('bakin_exec_submit_step', `taskId=${task.id} stepId=${stepContext.stepId} --args '<json output>'`)}`)
-  lines.push('')
-  lines.push(`# Log progress (mandatory, every major step)`)
-  lines.push(`${wfMc('bakin_exec_tasks_log_progress', `taskId=${task.id} message="<update>"`)}`)
-  lines.push('')
-  lines.push(`# Check your current step details if needed`)
-  lines.push(`${wfMc('bakin_exec_get_step', `taskId=${task.id}`)}`)
-  lines.push('')
-  lines.push('# --- Execution tools for doing actual work ---')
-  lines.push('')
-  // Channel posting only for output/publish steps (others have "NO SIDE EFFECTS")
-  lines.push(...sharedExecutionToolDocs(agentName, task.id, { allowChannelPost: stepContext.type === 'output' }))
-  lines.push('```')
-  lines.push('')
-
-  // ─── Stop Instruction ───────────────────────────────────────────────
-  lines.push('## AFTER SUBMITTING')
-  lines.push('')
-  lines.push('After bakin_exec_submit_step returns success, your work is done. Do NOT:')
-  lines.push('- Generate additional outputs or deliverables')
-  lines.push('- Start work on what you think the next step might be')
-  lines.push('- Send messages about what should happen next')
-  lines.push('- Move the task to Done (the workflow engine handles this)')
-
-  return lines.join('\n')
-}
-
-export function getDispatchInfo(contentDir: string): Record<string, unknown> {
-  const settings = getSettings()
-  const state = loadDispatchState(contentDir)
-  const now = Date.now()
-  const interval = settings.dispatch.intervalMs
-  const baseline = state.lastRun || state.serverStart || now
-
-  // Calculate the next run time that's actually in the future.
-  // If we've missed one or more intervals (e.g. dispatch was stuck),
-  // advance to the next upcoming tick rather than showing 0:00.
-  let nextRun = baseline + interval
-  if (nextRun <= now) {
-    const elapsed = now - baseline
-    const missedIntervals = Math.floor(elapsed / interval)
-    nextRun = baseline + (missedIntervals + 1) * interval
-  }
-  const secondsUntilNext = Math.max(0, Math.round((nextRun - now) / 1000))
-
-  return {
-    intervalMs: interval,
-    intervalMin: interval / 60000,
-    lastRun: state.lastRun ? new Date(state.lastRun).toISOString() : null,
-    nextRun: new Date(nextRun).toISOString(),
-    secondsUntilNext,
-    dispatching,
-    dispatchedCount: state.dispatched.length,
-  }
-}
+} from './dispatch-types'
